+++ HISTORY.md
@@ -1,17 +1,22 @@
 
 # Release history
 
-### main branch
+### Aider v0.48.0
 
+- Performance improvements for large/mono repos.
 - Added `--subtree-only` to limit aider to current directory subtree.
   - Should help with large/mono repo performance.
 - New `/add-clipboard-image` to add images to the chat from your clipboard.
+- Use `--map-tokens 1024` to use repo map with any model.
 - Support for Sonnet's 8k output window.
   - [Aider already supported infinite output from Sonnet.](https://aider.chat/2024/07/01/sonnet-not-lazy.html)
-- Performance improvements for large repos.
 - Workaround litellm bug for retrying API server errors.
 - Upgraded dependencies, to pick up litellm bug fixes.
-- Aider wrote 48% of the code since the last release.
+- Aider wrote 44% of the code in this release.
+
+### Aider v0.47.1
+
+- Improvements to conventional commits prompting.
 
 ### Aider v0.47.0